src/requirements.txt
Django==2.1.1
pytz==2018.5
mysqlclient==1.3.12
django-admin-rangefilter==0.3.6